requirements.txt
tensorflow
opencv-python-headless
numpy
streamlit
Pillow